make Talker usable with Mobile Safari
This ensures the chat box is alwas at the bottom of the page. Used to
be placed in the middle of the screen making the user scroll up
and down to talk and read the conversation. Some things should be improved
for Mobile Safari, but at least now it's usable.